Update adding-structure-to-smart-contracts.md
In the documentation, we speak directly to the user (in this case a designer or developer of CorDapps). So, much of these changes are just focusing on the user (you) and not on us (we). 

I've also tried to thin down the opening section and get to the benefits of CDL in terms of structure.